Move to_copy IR to to_copy.h
